+++ app/(blog)/sitemaps/layout.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Header from "@/app/components/Header";
+import Footer from "@/app/components/Footer";
 
 export const metadata: Metadata = {
   title: "サイトマップ | 編集箇所サイト名",
